# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Route {
  id: string;
  name: string;
  from: string;
  to: string;
  mode: string;
  durationMinutes: number;
  fare: number;
  status: string;
  nextDeparture?: string;
  color?: string;
}

export interface TripOption {
  id: string;
  routeName: string;
  mode: string;
  durationMinutes: number;
  fare: number;
  departure: string;
  arrival: string;
  occupancy: number;
  stops?: number;
  status?: string;
}

export interface ActiveTrip {
  id: string;
  routeName: string;
  destination: string;
  etaMinutes: number;
  progress: number;
  status: string;
  vehicle?: string;
}

export interface RecentTrip {
  id: string;
  routeName: string;
  date: string;
  fare: number;
  status: string;
}

export interface Impact {
  minutesSaved: number;
  carbonReduced: number;
  tripsCompleted: number;
}

export interface CommuterSummary {
  firstName: string;
  walletBalance: number;
  activeTrip: ActiveTrip | null;
  recentTrips: RecentTrip[];
  savedRoutes: Route[];
  impact: Impact;
}

export interface BookingInput {
  tripId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  passengerCount: number;
  paymentMethod: string;
}

export interface Booking {
  id: string;
  tripId: string;
  routeName: string;
  passengerCount: number;
  totalFare: number;
  paymentMethod: string;
  status: string;
  createdAt: string;
}

export interface PaymentMethod {
  id: string;
  label: string;
  type: string;
  isDefault: boolean;
}

export interface Transaction {
  id: string;
  description: string;
  date: string;
  amount: number;
  status: string;
}

export interface Wallet {
  balance: number;
  currency: string;
  autoTopup: boolean;
  methods: PaymentMethod[];
  transactions: Transaction[];
}

export interface ReportInput {
  reportType: string;
  /** @minLength 3 */
  title: string;
  /** @minLength 5 */
  description: string;
}

export interface Report {
  id: string;
  reportType: string;
  title: string;
  description: string;
  status: string;
  createdAt: string;
}

export type ListRoutesParams = {
search?: string;
};

export type SearchTripsParams = {
origin: string;
destination: string;
};