# lib/api-zod/src/generated/types/activeTrip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveTrip {
  id: string;
  routeName: string;
  destination: string;
  etaMinutes: number;
  progress: number;
  status: string;
  vehicle?: string;
}