# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: string;
  tripId: string;
  routeName: string;
  passengerCount: number;
  totalFare: number;
  paymentMethod: string;
  status: string;
  createdAt: string;
}