# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  tripId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  passengerCount: number;
  paymentMethod: string;
}