# lib/api-zod/src/generated/types/commuterSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActiveTrip } from './activeTrip';
import type { Impact } from './impact';
import type { RecentTrip } from './recentTrip';
import type { Route } from './route';

export interface CommuterSummary {
  firstName: string;
  walletBalance: number;
  activeTrip: ActiveTrip | null;
  recentTrips: RecentTrip[];
  savedRoutes: Route[];
  impact: Impact;
}