# lib/api-zod/src/generated/types/impact.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Impact {
  minutesSaved: number;
  carbonReduced: number;
  tripsCompleted: number;
}