# lib/api-zod/src/generated/types/recentTrip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentTrip {
  id: string;
  routeName: string;
  date: string;
  fare: number;
  status: string;
}