# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Report {
  id: string;
  reportType: string;
  title: string;
  description: string;
  status: string;
  createdAt: string;
}