# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  reportType: string;
  /** @minLength 3 */
  title: string;
  /** @minLength 5 */
  description: string;
}