# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Route {
  id: string;
  name: string;
  from: string;
  to: string;
  mode: string;
  durationMinutes: number;
  fare: number;
  status: string;
  nextDeparture?: string;
  color?: string;
}