# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: string;
  description: string;
  date: string;
  amount: number;
  status: string;
}