# lib/api-zod/src/generated/types/tripOption.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TripOption {
  id: string;
  routeName: string;
  mode: string;
  durationMinutes: number;
  fare: number;
  departure: string;
  arrival: string;
  occupancy: number;
  stops?: number;
  status?: string;
}