# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from './paymentMethod';
import type { Transaction } from './transaction';

export interface Wallet {
  balance: number;
  currency: string;
  autoTopup: boolean;
  methods: PaymentMethod[];
  transactions: Transaction[];
}